Updated newIDE to react-scripts 1.0.1
