+++ newIDE/app/package.json
@@ -5,25 +5,19 @@
   "license": "MIT",
   "homepage": ".",
   "devDependencies": {
-    "babel-eslint": "6.1.2",
-    "eslint": "3.5.0",
-    "eslint-config-react-app": "0.2.1",
-    "eslint-plugin-flowtype": "2.18.1",
-    "eslint-plugin-import": "1.12.0",
-    "eslint-plugin-jsx-a11y": "2.2.2",
-    "eslint-plugin-react": "6.3.0",
     "flow-bin": "^0.46.0",
     "prettier": "^0.22.0",
-    "react-scripts": "0.6.1",
-    "shelljs": "^0.7.7",
-    "@storybook/react": "^2.21.0"
+    "react-scripts": "1.0.1",
+    "shelljs": "^0.7.7"
   },
   "dependencies": {
+    "@storybook/react": "^3.0.0-alpha.0",
     "flat": "2.0.1",
     "lodash": "^4.16.4",
     "material-ui": "0.17.3",
     "pixi-simple-gesture": "0.2.2",
     "pixi.js": "3.0.11",
+    "prop-types": "^15.5.10",
     "react": "15.4.2",
     "react-addons-css-transition-group": "15.4.2",
     "react-addons-perf": "15.4.2",
