+++ newIDE/app/src/InstancesEditor/SelectionRectangle.js
@@ -1,5 +1,5 @@
-const gd = global.gd;
 import PIXI from 'pixi.js';
+const gd = global.gd;
 
 export default class SelectionRectangle {
   constructor(
